backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
psycopg[binary]==3.2.10
python-dotenv==1.1.1
gunicorn==23.0.0
Pillow==11.2.1
